Registration, Login, Add Product, Cart
In registration, I'm verifying the user based on the provided Email by sending a verification link to Email. When the verification of Email is done, only then the user gets logged into the system. For adding products, we are uploading an excel sheet. Using this sheet shopkeeper owners can download the list and update as per their product list. In card we can add products, update and delete products.